separated body & proto content files, clean up front-end, everything fully functional with dummy data
